requirements.txt
numpy 
pandas
matplotlib
scikit-learn
tensorflow-cpu
opencv_python